Change return type of calculateWeight() to double, to match X and Y coordinate types.

+++ tsp-ga/weighting.c
@@ -8,7 +8,7 @@
 #include "weighting.h"
 
 
-float calculateWeight(location *loc1, location *loc2)
+double calculateWeight(location *loc1, location *loc2)
 {
